Refactor main.py to separate out Health and Admin endpoints
