+++ src/fidesapi/main.py
@@ -1,13 +1,10 @@
 """
 Contains the code that sets up the API.
 """
-
-
 from datetime import datetime
-from enum import Enum
 from logging import WARNING
 from pathlib import Path
-from typing import Callable, Dict
+from typing import Callable
 
 from fastapi import FastAPI, HTTPException, Request, Response, status
 from fastapi.responses import FileResponse
@@ -15,13 +12,10 @@ from fastapi.staticfiles import StaticFiles
 from loguru import logger as log
 from uvicorn import Config, Server
 
-import fidesctl
 from fidesapi import view
-from fidesapi.database import database
-from fidesapi.database.database import get_db_health
-from fidesapi.routes import crud, datamap, generate, validate, visualize
+from fidesapi.database.database import configure_db
+from fidesapi.routes import admin, crud, datamap, generate, health, validate, visualize
 from fidesapi.routes.util import API_PREFIX, WEBAPP_DIRECTORY, WEBAPP_INDEX
-from fidesapi.utils.errors import get_full_exception_name
 from fidesapi.utils.logger import setup as setup_logging
 from fidesctl.core.config import FidesctlConfig, get_config
 
@@ -40,20 +34,12 @@ def configure_routes() -> None:
     app.include_router(generate.router)
     app.include_router(datamap.router)
     app.include_router(validate.router)
+    app.include_router(admin.router)
+    app.include_router(health.router)
 
 
 # Configure the routes here so we can generate the openapi json file
 configure_routes()
-
-
-async def configure_db(database_url: str) -> None:
-    "Set up the db to be used by the app."
-    try:
-        database.create_db_if_not_exists(database_url)
-        await database.init_db(database_url)
-    except Exception as error:  # pylint: disable=broad-except
-        error_type = get_full_exception_name(error)
-        log.error(f"Unable to configure database: {error_type}: {error}")
 
 
 @app.on_event("startup")
@@ -94,74 +80,6 @@ async def log_request(request: Request, call_next: Callable) -> Response:
         path=request.url.path,
     ).info("Request received")
     return response
-
-
-@app.get(
-    f"{API_PREFIX}/health",
-    response_model=Dict[str, str],
-    responses={
-        status.HTTP_200_OK: {
-            "content": {
-                "application/json": {
-                    "example": {
-                        "status": "healthy",
-                        "version": "1.0.0",
-                        "database": "healthy",
-                    }
-                }
-            }
-        },
-        status.HTTP_503_SERVICE_UNAVAILABLE: {
-            "content": {
-                "application/json": {
-                    "example": {
-                        "detail": {
-                            "status": "healthy",
-                            "version": "1.0.0",
-                            "database": "unhealthy",
-                        }
-                    }
-                }
-            }
-        },
-    },
-    tags=["Health"],
-)
-async def health() -> Dict:
-    "Confirm that the API is running and healthy."
-    database_health = get_db_health(CONFIG.api.sync_database_url)
-    response = {
-        "status": "healthy",
-        "version": str(fidesctl.__version__),
-        "database": database_health,
-    }
-
-    for key in response:
-        if response[key] == "unhealthy":
-            raise HTTPException(
-                status_code=status.HTTP_503_SERVICE_UNAVAILABLE, detail=response
-            )
-
-    return response
-
-
-class DBActions(str, Enum):
-    "The available path parameters for the `/admin/db/{action}` endpoint."
-    init = "init"
-    reset = "reset"
-
-
-@app.post(API_PREFIX + "/admin/db/{action}", tags=["Admin"])
-async def db_action(action: DBActions) -> Dict:
-    """
-    Initiate one of the enumerated DBActions.
-    """
-    action_text = "initialized"
-    if action == DBActions.reset:
-        database.reset_db(CONFIG.api.sync_database_url)
-        action_text = DBActions.reset
-    await configure_db(CONFIG.api.sync_database_url)
-    return {"data": {"message": f"Fidesctl database {action_text}"}}
 
 
 # Configure the static file paths last since otherwise it will take over all paths
